% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{create_boot_samps}
\alias{create_boot_samps}
\title{Generate Bootstrap Resamples from Count Data}
\usage{
create_boot_samps(t, y, n, num_boot)
}
\arguments{
\item{t}{A numeric vector of ages (or identifiers for age groups).}

\item{y}{A numeric vector of seropositive counts for each age group.}

\item{n}{A numeric vector of total sample sizes for each age group.}

\item{num_boot}{Number of bootstrap resamples to generate.}
}
\value{
A list of length \code{num_boot}. Each element is itself a list with:
\describe{
\item{\code{t}}{Vector of unique age values in the bootstrap sample.}
\item{\code{y}}{Bootstrap seropositive counts.}
\item{\code{n}}{Bootstrap total sample sizes.}
}
}
\description{
Produces a list of bootstrap resamples by sampling with replacement from the pseudo-individual-level data.
Each resample is returned in aggregated count form.
}
\examples{
t <- c(1, 2, 3)
y <- c(2, 3, 4)
n <- c(4, 5, 6)
create_boot_samps(t, y, n, num_boot = 100)

}
