% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VynnyckyD.R
\name{VynneckyD}
\alias{VynneckyD}
\title{Vynnycky Catalytic Model D (Constant FOI, Perfect Assay Sensitivity)}
\usage{
VynneckyD(t, y, n)
}
\arguments{
\item{t}{A numeric vector of age groups.}

\item{y}{A numeric vector of seropositive counts corresponding to each age group.}

\item{n}{A numeric vector of total sample sizes corresponding to each age group.}
}
\value{
A list with the following components:
\describe{
\item{\code{par}}{Named list with the maximum likelihood estimate for \code{foi} (force of infection).}
\item{\code{CIs}}{Named list with the 95\\% bootstrap confidence interval for \code{foi}.}
\item{\code{boot_params}}{Named list containing the bootstrap samples for \code{foi}.}
}
}
\description{
Fits a catalytic model assuming a constant force of infection (FOI) across all ages and perfect assay sensitivity (\eqn{ρ = 1}).
}
\details{
The seroprevalence function is modeled as:
\deqn{
\pi(t) = 1 - \exp(-\lambda \cdot (t - 0.5))
}

This model assumes that all past infections are perfectly detected and estimates a single FOI parameter (\eqn{\lambda}).
}
