% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VynnyckyA.R
\name{VynnyckyA}
\alias{VynnyckyA}
\title{Vynnycky Catalytic Model A}
\usage{
VynnyckyA(t, y, n)
}
\arguments{
\item{t}{A numeric vector of age groups.}

\item{y}{A numeric vector of seropositive counts corresponding to each age group.}

\item{n}{A numeric vector of total sample sizes corresponding to each age group.}
}
\value{
A list with the following components:
\describe{
\item{\code{par}}{Named list of MLE estimates for \code{rho}, \code{foi_y}, and \code{foi_o}.}
\item{\code{CIs}}{Named list of 95\% bootstrap confidence intervals for each parameter.}
\item{\code{boot_params}}{Named list of bootstrap samples for each parameter. Useful for diagnostics or visualisation.}
}
}
\description{
Fits a piecewise catalytic model of seroprevalence that accounts for different forces of infection
in younger (<13 years) and older (≥13 years) individuals, and includes an unknown assay sensitivity parameter.
}
\details{
The model assumes:
\deqn{π(t) = ρ(1 - e^{-λ_y (t - 0.5)})} for \eqn{t < 13}
\deqn{π(t) = ρ(1 - e^{-λ_y (12.5) - λ_o (t - 13)})} for \eqn{t \geq 13}
where:
\itemize{
\item \eqn{ρ} is the assay sensitivity (maximum achievable seroprevalence),
\item \eqn{λ_y} is the force of infection for individuals under 13,
\item \eqn{λ_o} is the force of infection for individuals 13 and older.
}
}
