% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VynnyckyB.R
\name{VynnyckyB}
\alias{VynnyckyB}
\title{Vynnycky Catalytic Model B (Piecewise Constant FOI, Assay Sensitivity = 1)}
\usage{
VynnyckyB(t, y, n)
}
\arguments{
\item{t}{A numeric vector of age groups.}

\item{y}{A numeric vector of seropositive counts corresponding to each age group.}

\item{n}{A numeric vector of total sample sizes corresponding to each age group.}
}
\value{
A list with the following elements:
\describe{
\item{\code{par}}{Named list of maximum likelihood estimates: \code{foi_y}, \code{foi_o}.}
\item{\code{CIs}}{Named list of 95\% bootstrap confidence intervals for each parameter.}
\item{\code{boot_params}}{Named list containing the bootstrap samples for \code{foi_y} and \code{foi_o}.}
}
}
\description{
Fits a piecewise constant force of infection (FOI) catalytic model assuming full assay sensitivity (\eqn{ρ = 1}).
}
\details{
The seroprevalence function is modeled as:
\deqn{
\pi(t) =
\begin{cases}
  1 - \exp(-\lambda_y \cdot (t - 0.5)), & \text{for } t < 13 \\
  1 - \exp(-\lambda_y \cdot 12.5 - \lambda_o \cdot (t - 13)), & \text{for } t \geq 13
\end{cases}
}

This structure captures different FOI values for children (\eqn{\lambda_y}) and older individuals (\eqn{\lambda_o}).
}
