% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuenchGeneral.R
\name{MuenchGeneral}
\alias{MuenchGeneral}
\title{Muench's Catalytic Model}
\usage{
MuenchGeneral(t, y, n)
}
\arguments{
\item{t}{Vector of age groups}

\item{y}{Seropositive counts}

\item{n}{Sample sizes}
}
\value{
A list with MLE estimates, CIs, and bootstrap samples
}
\description{
Fits the catalytic model: π(t) = k(l - e^{-λt})
}
